+++ src/main/java/com/nisovin/shopkeepers/shopobjects/virtual/SKVirtualShopObject.java
@@ -63,11 +63,16 @@ public class SKVirtualShopObject extends AbstractShopObject implements VirtualSh
 		return null;
 	}
 
+	@Override
+	public Location getTickVisualizationParticleLocation() {
+		return null;
+	}
+
 	// TICKING
 
 	@Override
 	public void tick() {
-		// Nothing to do
+		super.tick();
 	}
 
 	// NAMING